fix(EMS-1951): rename complete application cypress command file names
