+++ e2e-tests/commands/insurance/complete-prepare-application-section-single-policy-type.js
